FileAnalyzer/requirements.txt
Flask
Authlib
python-dotenv
Flask-SQLAlchemy
psycopg2-binary
gunicorn
requests